move logging time latch to common
