+++ beacon_node/network/Cargo.toml
@@ -11,7 +11,6 @@ matches = "0.1.8"
 exit-future = "0.2.0"
 slog-term = "2.6.0"
 slog-async = "2.5.0"
-logging = { path = "../../common/logging" }
 environment = { path = "../../lighthouse/environment" }
 
 [dependencies]
@@ -35,6 +34,7 @@ fnv = "1.0.7"
 rlp = "0.5.0"
 lazy_static = "1.4.0"
 lighthouse_metrics = { path = "../../common/lighthouse_metrics" }
+logging = { path = "../../common/logging" }
 task_executor = { path = "../../common/task_executor" }
 igd = "0.11.1"
 itertools = "0.10.0"
